% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{create_python_env}
\alias{create_python_env}
\title{Create and configure Python environment}
\usage{
create_python_env(env_name = "r-py-env", python_version = "3.9")
}
\arguments{
\item{env_name}{Character. Name of the virtual environment to create}

\item{python_version}{Character. Python version to use (e.g., "3.8")}
}
\value{
Logical indicating success
}
\description{
Create and configure Python environment
}
\keyword{internal}
