% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{initialize_python_env}
\alias{initialize_python_env}
\title{Initialize or repair Python environment}
\usage{
initialize_python_env(force = FALSE)
}
\arguments{
\item{force}{Logical. If TRUE, recreates the environment even if it exists}
}
\description{
Initialize or repair Python environment
}
